Move MachineAppLXDProfileWatcher and tests to new files.
